I have started to make input and output.

+++ src/common.h
@@ -13,8 +13,10 @@
 
 #include <algorithm>
 
-typedef int64_t sh_time_t;
-#define sh_time_max LONG_MAX
+#include <tinyxml2.h>
+
+typedef uint32_t sh_time_t;
+#define sh_time_max INT_MAX
 #define sh_time_min 0
 #define sh_time_look_ahead 1000
 #define sh_time_window (sh_time_look_ahead / 5)
